tdd/binary_search_tree.py
class TreeNode:
    def __init__(self, key, value, left=None, right=None, parent=None):
        self.key = key
        self.value = value
        self.left_child = left
        self.right_child = right
        self.parent = parent

class BinarySearchTree:
    def __init__(self):
        self.root = None
        self.size = 0

    def __len__(self):
        return self.size

    def put(self, key, value):
        if self.root:
            self._put(key, value, self.root)
        else:
            self.root = TreeNode(key, value)
        self.size += 1

    def _put(self, key, value, current_node):
        if key < current_node.key:
            if current_node.left_child:
                self._put(key, value, current_node.left_child)
            else:
                current_node.left_child = TreeNode(key, value)
        else:
            if current_node.right_child:
               self._put(key, value, current_node.right_child)
            else:
                current_node.right_child = TreeNode(key, value)

    def get(self, key):
        return self._get(key, self.root)

    def _get(self, key, current_node):
        if not current_node:
            return None
        elif key == current_node.key:
            return current_node.value
        elif key < current_node.key:
            return self._get(key, current_node.left_child)
        else:
            return self._get(key, current_node.right_child)